health: Remove unrequired uint64 conversation

+++ health/check_peers.go
@@ -11,12 +11,12 @@ var (
 )
 
 // checkMinPeers returns 'errNotEnoughPeers' if the current peer count its lower than 'minPeerCount'
-func checkMinPeers(ec ethClient, minPeerCount uint) error {
+func checkMinPeers(ec ethClient, minPeerCount uint64) error {
 	peerCount, err := ec.PeerCount(context.TODO())
 	if err != nil {
 		return err
 	}
-	if uint64(peerCount) < uint64(minPeerCount) {
+	if peerCount < minPeerCount {
 		return fmt.Errorf("%w: %d (minimum %d)", errNotEnoughPeers, peerCount, minPeerCount)
 	}
 	return nil
